refactor: consolidate shift type form
