+++ app/src/utils/generateId.ts
@@ -1,1 +1,29 @@
-export const generateId = () => Math.floor(Math.random() * 1000000 + 1);
+import { RowDataPacket } from "mysql2";
+
+import { pool } from "lib/database";
+
+const randomNum = () => Math.floor(Math.random() * 1000000 + 1);
+
+export const checkIsIdExists = async (query: string, id: number) => {
+  const [dbIdList] = await pool.query<RowDataPacket[]>(query, [id]);
+  const dbIdFirst = dbIdList[0];
+
+  return Boolean(dbIdFirst);
+};
+
+export const generateId = (query: string) => {
+  let idNew = 0;
+
+  const changeNum = async () => {
+    idNew = randomNum();
+
+    const isIdExists = await checkIsIdExists(query, idNew);
+
+    // if shift name ID exists already
+    // then execute function recursively
+    if (isIdExists) changeNum();
+  };
+  changeNum();
+
+  return idNew;
+};
